Added Convert-StaticIP-to-DHCP-Alternate.ps1
This powershell script requires elevated privileges.

When run, it will take the current static IP configuration of a connected network interface, and move that to the Alternate Configuration. Then it will make the default configuration DHCP.

This is useful when preparing to move a moderately sized network of client machines to DHCP, but don't want to manually change all the settings. Applying this as a startup script for Group Policy, for example, can speed up the deployment.